refactor: consolidate start_redis_container into docker.py, remove duplicate _default_start_redis_container
